create gx subpackage and relocate gx files
